Refactor appointment repository methods to introduce findUserAppointmentsWithFilters and streamline user appointment retrieval in userService
